move finder code to share with all content
