+++ src/content/indicator.js
